move mps_create_file_lists.py to tkal_create_file_lists.py
